+++ src/transaction-signer.js
